Code cleanup and move stuff around
